azure-functions-python-try/pyproject.toml
[project]
name = "azure-functions-python-try"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-functions>=1.24.0",
]
